+++ airbyte-db/lib/src/test/java/io/airbyte/db/jdbc/TestJdbcUtils.java
@@ -65,6 +65,7 @@ public class TestJdbcUtils {
   private static PostgreSQLContainer<?> PSQL_DB;
 
   private BasicDataSource dataSource;
+  private static final JdbcSourceOperations sourceOperations = JdbcUtils.getDefaultSourceOperations();
 
   @BeforeAll
   static void init() {
@@ -115,7 +116,7 @@ public class TestJdbcUtils {
     try (final Connection connection = dataSource.getConnection()) {
       final ResultSet rs = connection.createStatement().executeQuery("SELECT * FROM id_and_name;");
       rs.next();
-      assertEquals(RECORDS_AS_JSON.get(0), JdbcUtils.rowToJson(rs));
+      assertEquals(RECORDS_AS_JSON.get(0), sourceOperations.rowToJson(rs));
     }
   }
 
@@ -123,7 +124,7 @@ public class TestJdbcUtils {
   void testToStream() throws SQLException {
     try (final Connection connection = dataSource.getConnection()) {
       final ResultSet rs = connection.createStatement().executeQuery("SELECT * FROM id_and_name;");
-      final List<JsonNode> actual = JdbcUtils.toStream(rs, JdbcUtils::rowToJson).collect(Collectors.toList());
+      final List<JsonNode> actual = sourceOperations.toStream(rs, sourceOperations::rowToJson).collect(Collectors.toList());
       assertEquals(RECORDS_AS_JSON, actual);
     }
   }
@@ -149,21 +150,21 @@ public class TestJdbcUtils {
 
       // insert the bit here to stay consistent even though setStatementField does not support it yet.
       ps.setString(1, "1");
-      JdbcUtils.setStatementField(ps, 2, JDBCType.BOOLEAN, "true");
-      JdbcUtils.setStatementField(ps, 3, JDBCType.SMALLINT, "1");
-      JdbcUtils.setStatementField(ps, 4, JDBCType.INTEGER, "1");
-      JdbcUtils.setStatementField(ps, 5, JDBCType.BIGINT, "1");
-      JdbcUtils.setStatementField(ps, 6, JDBCType.FLOAT, "1.0");
-      JdbcUtils.setStatementField(ps, 7, JDBCType.DOUBLE, "1.0");
-      JdbcUtils.setStatementField(ps, 8, JDBCType.REAL, "1.0");
-      JdbcUtils.setStatementField(ps, 9, JDBCType.NUMERIC, "1");
-      JdbcUtils.setStatementField(ps, 10, JDBCType.DECIMAL, "1");
-      JdbcUtils.setStatementField(ps, 11, JDBCType.CHAR, "a");
-      JdbcUtils.setStatementField(ps, 12, JDBCType.VARCHAR, "a");
-      JdbcUtils.setStatementField(ps, 13, JDBCType.DATE, "2020-11-01T00:00:00Z");
-      JdbcUtils.setStatementField(ps, 14, JDBCType.TIME, "1970-01-01T05:00:00Z");
-      JdbcUtils.setStatementField(ps, 15, JDBCType.TIMESTAMP, "2001-09-29T03:00:00Z");
-      JdbcUtils.setStatementField(ps, 16, JDBCType.BINARY, "61616161");
+      sourceOperations.setStatementField(ps, 2, JDBCType.BOOLEAN, "true");
+      sourceOperations.setStatementField(ps, 3, JDBCType.SMALLINT, "1");
+      sourceOperations.setStatementField(ps, 4, JDBCType.INTEGER, "1");
+      sourceOperations.setStatementField(ps, 5, JDBCType.BIGINT, "1");
+      sourceOperations.setStatementField(ps, 6, JDBCType.FLOAT, "1.0");
+      sourceOperations.setStatementField(ps, 7, JDBCType.DOUBLE, "1.0");
+      sourceOperations.setStatementField(ps, 8, JDBCType.REAL, "1.0");
+      sourceOperations.setStatementField(ps, 9, JDBCType.NUMERIC, "1");
+      sourceOperations.setStatementField(ps, 10, JDBCType.DECIMAL, "1");
+      sourceOperations.setStatementField(ps, 11, JDBCType.CHAR, "a");
+      sourceOperations.setStatementField(ps, 12, JDBCType.VARCHAR, "a");
+      sourceOperations.setStatementField(ps, 13, JDBCType.DATE, "2020-11-01T00:00:00Z");
+      sourceOperations.setStatementField(ps, 14, JDBCType.TIME, "1970-01-01T05:00:00Z");
+      sourceOperations.setStatementField(ps, 15, JDBCType.TIMESTAMP, "2001-09-29T03:00:00Z");
+      sourceOperations.setStatementField(ps, 16, JDBCType.BINARY, "61616161");
 
       ps.execute();
 
@@ -238,7 +239,7 @@ public class TestJdbcUtils {
     final ResultSet resultSet = connection.createStatement().executeQuery("SELECT * FROM data;");
 
     resultSet.next();
-    final JsonNode actual = JdbcUtils.rowToJson(resultSet);
+    final JsonNode actual = sourceOperations.rowToJson(resultSet);
 
     final ObjectNode expected = (ObjectNode) Jsons.jsonNode(Collections.emptyMap());
     expected.put("bit", true);
@@ -272,7 +273,7 @@ public class TestJdbcUtils {
     final int columnCount = resultSet.getMetaData().getColumnCount();
     final Map<String, JsonSchemaPrimitive> actual = new HashMap<>(columnCount);
     for (int i = 1; i <= columnCount; i++) {
-      actual.put(resultSet.getMetaData().getColumnName(i), JdbcUtils.getType(JDBCType.valueOf(resultSet.getMetaData().getColumnType(i))));
+      actual.put(resultSet.getMetaData().getColumnName(i), sourceOperations.getType(JDBCType.valueOf(resultSet.getMetaData().getColumnType(i))));
     }
 
     final Map<String, JsonSchemaPrimitive> expected = ImmutableMap.<String, JsonSchemaPrimitive>builder()
