+++ airbyte-integrations/connectors/destination-jdbc/src/test/java/io/airbyte/integrations/destination/jdbc/SqlOperationsUtilsTest.java
@@ -106,7 +106,7 @@ class SqlOperationsUtilsTest {
 
     final List<JsonNode> actualRecords = database.bufferedResultSetQuery(
         connection -> connection.createStatement().executeQuery("SELECT * FROM RIVERS"),
-        JdbcUtils::rowToJson);
+        JdbcUtils.getDefaultSourceOperations()::rowToJson);
 
     final List<JsonNode> expectedRecords = Lists.newArrayList(
         Jsons.jsonNode(ImmutableMap.builder()
